major refactor for the structure of the project including moving types and interfaces into the types file and importing them when needed, also any component that shows up when a page is loaded and the views were removed from their respective pages and moved into the views folder to help with clarity. Checked and tested all the imports and components to work correctly.
