+++ components/views/PublicProfileView.tsx
@@ -1,0 +1,88 @@
+import React from 'react';
+import { View, Text, Image, StyleSheet, ActivityIndicator } from 'react-native';
+import { PublicProfileViewProps } from '../../types/types';
+
+const PublicProfileView: React.FC<PublicProfileViewProps> = ({ user, loading }) => {
+  if (loading) {
+    return (
+      <View style={styles.center}>
+        <ActivityIndicator size="large" color="#333" />
+      </View>
+    );
+  }
+
+  return (
+    <View style={styles.container}>
+      <View style={styles.card}>
+        <Image
+          source={{
+            uri: 'https://i.pinimg.com/736x/2f/15/f2/2f15f2e8c688b3120d3d26467b06330c.jpg',
+          }}
+          style={styles.profileImage}
+        />
+
+        <Text style={styles.name}>
+          {user.username || 'No username'}
+        </Text>
+        <Text style={styles.email}>
+          {user.email || 'No email'}
+        </Text>
+        <Text style={styles.detail}>
+          Dorm: {user.location ?? 'N/A'} / Floor: {user.floor ?? 'N/A'}
+        </Text>
+      </View>
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  center: {
+    flex: 1,
+    justifyContent: 'center',
+  },
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+    paddingHorizontal: 10,
+    backgroundColor: '#90BE6D',
+  },
+  card: {
+    width: '90%',
+    backgroundColor: '#FFE2D1',
+    borderRadius: 16,
+    padding: 30,
+    alignItems: 'center',
+    shadowColor: '#000',
+    shadowOffset: { width: 0, height: 4 },
+    shadowOpacity: 0.2,
+    shadowRadius: 8,
+  },
+  profileImage: {
+    width: 120,
+    height: 120,
+    borderRadius: 60,
+    marginBottom: 10,
+    alignSelf: 'center',
+  },
+  name: {
+    fontSize: 22,
+    fontWeight: 'bold',
+    marginVertical: 5,
+    textAlign: 'center',
+  },
+  email: {
+    fontSize: 16,
+    marginBottom: 10,
+    color: '#666',
+    textAlign: 'center',
+  },
+  detail: {
+    fontSize: 16,
+    color: '#555',
+    marginBottom: 15,
+    textAlign: 'center',
+  },
+});
+
+export default PublicProfileView;
